// src/types/blog-category.ts
interface BlogCategory {
    id: number;
    name: string;
    name_en: string;
    logo: string;
    image: string;
    status: number;
    seo_title: string;
    seo_keywords: string;
    seo_description: string;
    created_at: string;
    updated_at: string;
    blogs_count: number;
}

export default BlogCategory